+++ tests/ndarray/test_persistency.py
@@ -35,9 +35,7 @@ def test_persistency(shape, chunks, blocks, urlpath, contiguous, dtype):
 
     size = int(np.prod(shape))
     nparray = np.arange(size, dtype=dtype).reshape(shape)
-    _ = blosc2.asarray(
-        nparray, chunks=chunks, blocks=blocks, dtype=dtype, urlpath=urlpath, contiguous=contiguous
-    )
+    _ = blosc2.asarray(nparray, chunks=chunks, blocks=blocks, urlpath=urlpath, contiguous=contiguous)
     b = blosc2.open(urlpath)
 
     bc = b[:]
